Updated Readme.md with an autoclass example
Hi there, 

I wrote a library that fits quite nicely with enforce and would like to share it with you : [autoclass](https://github.com/smarie/python-autoclass). Notice how it completes enforce with validators and code generation.
If you find this work interesting, would you be so kind to add a reference to it in the readme of enforce? This pull request is an attempt to provide such an example in the readme but of course it is probably not the way you would put it so feel free to modify..

Thanks again for this great lib!
Sylvain